revision

+++ C++/helloworld.cpp
@@ -4,7 +4,7 @@ int main () {
 
     int x = 8 ;// store only absolute value
 
-    double days = 7.5; //store including integer
+    double days = 7.5; //store including decimal
     double price  = 9.99;
     double temperature = 18.8;
